Force version prefix
When cliquet-fxa is enabled via configuration (includes in .ini),
the cliquet initialization has not happened yet.

There the version prefix is not setup.

Currently, this means that cliquet-fxa views are accessible without
version prefix. Which leads to ugly code in client to reach them.

This PR brings the opposite approach. The views are only
reachable with version prefix. Which is what will happen in most
cases, and at Mozilla.